archive.c: add support for --recurse-submodules[=(all|checkedout)]
The --recurse-submodules option uses the enhanced read_tree_recursive()
to enable inclusion of submodules in the generated archive.

When invoked with `--recurse-submodules=all` all gitlink entries will
be traversed, and when invoked with --recurse-submodules=checkedout
(the default option) only gitlink entries with a git repo (i.e.
checked out submodules) will be traversed.

When a gitlink has been selected for traversal, it is required that all
objects necessary to perform this traversal are available in either the
primary odb or through an alternate odb. To this end, git archive will
insert the object database of the selected gitlink (when checked out)
as an alternate odb, using the new function add_alt_odb(). And since
alternates now can be added after parsing of objects/info/alternates,
the error message in link_alt_odb_entry() has been updated to not
mention this file.

Signed-off-by: Lars Hjemli <hjemli@gmail.com>
Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>