Fix check_preimage() to use memset() for initialization of struct checkout
Every other code site uses memset() to initialize auto variables of type
struct checkout, lets do that here too to be on the safe side in case this
structure is expanded.

Also add a test in checkout_entry() to avoid using base_dir when it is set
to NULL (as checkout_paths() in checkout.c also does).

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>