Prepare unpack_trees() for recursive checkout of submodules
To be able to checkout submodule content the member "recurse_submodules"
is added to "struct unpack_trees_options". In "unpack_callback()" it is
used to initialize the recently introduced member with the same name in
"struct checkout".

When this flag is set verify_uptodate_1() will consider submodules whose
HEAD differs from the commit in the superproject as modified. Even if
they both match, it is checked if the checkout in the submodule would
overwrite modified files.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>