submodule: Demonstrate failure to add with auto/safecrlf
If 'core.autocrlf' and 'core.safecrlf' are both enabled then
'git submodule add' fails with an error such as

 fatal: LF would be replaced by CRLF in .gitmodules
 Failed to register submodule 'submod'

because it generates a '.gitmodules' file with LF newlines that are
rejected by 'git add' under this configuration.  Demonstrate this known
breakage with a new test in t7400-submodule-basic covering the case.

Signed-off-by: Brad King <brad.king@kitware.com>