diff: don't fail on invalid gitlink when ignoring submodules
When a gitlink in a submodule doesn't point into a valid git directory
running "git diff" fails with:
   "fatal: Not a git repository: <invalid git dir>"
While this is ok, it also happens when the "--ignore-submodules" option is
used, which should avoid this problem.

The reason is that match_stat_with_submodule() calls ce_match_stat()
before it evaluates the ignore settings, which calls read_gitfile()
further down the callchain which dies on the invalid gitlink.

Solve that by evaluating the submodule ignore settings first and skip the
ce_match_stat() call if the submodule is ignored by command line option
or config.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>