status: don't fail on invalid gitlink when ignoring submodules
When a gitlink in a submodule doesn't point into a valid git directory
running "git status" fails with:
   "fatal: Not a git repository: <invalid git dir>"
While this is ok, it also happens when the "--ignore-submodules" option is
used, which should avoid this problem.

The reason is that refresh_index() calls read_gitfile() through
ce_compare_gitlink() which dies on the invalid gitlink.

Solve that by adding REFRESH_IGNORE_SUBMODULES to the refresh options
if the submodule is ignored by command line option or config.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>