!!! PRELIMINARY !!! Teach commit the --ignore-submodules option
When the gitlink in a submodule points to an invalid git directory the
preparation of the commit message template fails with:
   "fatal: Not a git repository: <invalid git dir>"
While this is ok as it informs the user about a problem, "git commit"
doesn't know the "--ignore-submodules" option yet which would avoid this
problem.

Teach "git commit" to not include submodule changes in the commit message
template when the new "--ignore-submodules" option is used an add a test
for that.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>