!!! WIP !!! submodule: use submodule.<name>.update config from .gitmodules as fallback
TODO: tests

For almost all submodule settings it makes sense to let upstream provide
resonable defaults and allow them to be different for different commits.
The only exception to that rule is the submodule URL, which should always
be taken from .git/config to allow overriding URLs that e.g. do not exist
anymore and to let the user choose to use a fork of the upstream repo.

The update flag is the only exception to that rule, as it is copied on
submodule init. On the other hand a later submodule sync then only updates
the URL, but not the update setting.

Prepare to make the update setting behave like all other per-submodule
configurations (except for the URL for the reasons stated above) by
falling back to the setting from .gitmodules if it isn't set in the usual
config places. This reduces the trouble for users which later access a
submodule cloned with a newer git version from the future which will not
copy the update setting to .git/config on init anymore.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>