!!! WIP !!! Teach rebase the --[no-]recurse-submodules option
TODO: teach interactive, merge and am rebase the same option

When using this option 'git rebase' will automatically update the work
tree of all initialized submodules (so they match the SHA-1 recorded in
the superproject) in each rebase stop. This makes calling 'git submodule
update' on each stop and after it finished obsolete, which was tedious and
error prone. If the option is given it is stored in the recurse_submodule
file in the state directory so that later rebase steps can reuse it. But
this commit only adds the option without any functionality, that will be
added to unpack_trees() in subsequent commits.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>