Teach read-tree the --[no-]recurse-submodules option
This new option will allow the user to not only update the work tree of
the superproject but to also update the work tree of all initialized
submodules (so they match the SHA-1 recorded in the superproject) when
used together with -u. But this commit only adds the option without any
functionality, that will be added to unpack_trees() in subsequent commits.

gitmodules_config() had to be extended to cope with an empty index file,
as it is perfectly ok to call read-tree in that case.

Signed-off-by: Jens Lehmann <Jens.Lehmann@web.de>